Reorganize and simplify constructs
