+++ crates/cgp-dispatch/src/providers/field_matchers/mod.rs
@@ -1,9 +1,11 @@
 mod extract_field;
 mod extract_first_field;
 mod extract_handle;
+mod field_value;
 mod first_field_value;
 
 pub use extract_field::*;
 pub use extract_first_field::*;
 pub use extract_handle::*;
+pub use field_value::*;
 pub use first_field_value::*;
