+++ crates/cgp-handler/src/providers/mod.rs
@@ -1,5 +1,4 @@
 mod compose;
-mod field;
 mod pipe;
 mod promote;
 mod promote_all;
@@ -9,7 +8,6 @@ mod return_input;
 mod try_promote;
 
 pub use compose::*;
-pub use field::*;
 pub use pipe::*;
 pub use promote::*;
 pub use promote_all::*;
